added delete_excel_file to the config file to allow the user to delete the excel file after it has been emailed. Additional changes were made to the analysis_functions file to verify the config file has a value entered and delete the excel file only if the user requests it to be deleted

+++ homescraper/analysis_functions.py
@@ -647,6 +647,7 @@ def config_file_required_values_present(config):
         "expected_capx_monthly": lambda x: isinstance(x, (float)) and 0 <= x <= 0.25,
         "expected_management_monthly": lambda x: isinstance(x, (float)) and 0 <= x <= 0.25,
         "insurance_rate_yearly": lambda x: isinstance(x, (float)) and 0 <= x <= 0.25,
+        "delete_excel_file": lambda x: isinstance(x, bool),
         "send_emails": lambda x: isinstance(x, bool),
     }
     
@@ -740,24 +741,28 @@ def create_target_values_dictionary(config):
     return target_values  
 
 
-def delete_file(file_path):
-    """Function to delete a given file"""
-    # Try to delete the file
-    try:
-        os.remove(file_path)
-        print(f"File {file_path} has been successfully deleted.")
-        
-    # Handle issues if the file is not found
-    except FileNotFoundError:
-        print(f"File not found: {file_path}")
-        
-    # Handle issues if permission is needed to delete the file
-    except PermissionError as pe:
-        print(f"Permission error deleting file {file_path}: {pe}")
-        
-    # Handle any additional exceptions
-    except Exception as e:
-        print(f"Error deleting file {file_path}: {e}")
+def delete_file(config, file_path):
+    """Function to delete a given file if the user requests"""
+    # Determine if the user wants the file deleted
+    if config['delete_excel_file']:
+        # Try to delete the file
+        try:
+            os.remove(file_path)
+            print(f"File {file_path} has been successfully deleted.")
+            
+        # Handle issues if the file is not found
+        except FileNotFoundError:
+            print(f"File not found: {file_path}")
+            
+        # Handle issues if permission is needed to delete the file
+        except PermissionError as pe:
+            print(f"Permission error deleting file {file_path}: {pe}")
+            
+        # Handle any additional exceptions
+        except Exception as e:
+            print(f"Error deleting file {file_path}: {e}")
+            
+    return
 
 
 def email_config_file_required_values_present(config):
